Chore: Rename folder from pagenationbutton to paginationbutton
